requirements.txt
requests>=2.28.0
feedparser>=6.0.0
openai>=1.0.0
trafilatura>=1.6.0
readability-lxml>=0.8.1
